dump functionality done

+++ src/test/java/org/zooper/remosko/test/persistence/TestImportExport.java
@@ -1,11 +1,11 @@
 package org.zooper.remosko.test.persistence;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 
 import java.io.File;
 
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 import org.apache.log4j.Logger;
 import org.junit.After;
 import org.junit.Before;
@@ -47,6 +47,7 @@ public class TestImportExport extends TestAbstract {
 	public void testImportExportAll() {
 		log.info("Starting test");
 		File f = null;
+		File f2 = null;
 		try {
 			String pathFile = importExport.exportJson(false);
 			assertTrue(pathFile.contains("dump"));
@@ -55,11 +56,20 @@ public class TestImportExport extends TestAbstract {
 			assertTrue(f.exists());
 			
 			importExport.importJsonFile(pathFile, false);
+			
+			String pathFile2 = importExport.exportJson(false);
+			f2 = new File(pathFile2);
+			//assertEquals(FileUtils.readFileToString(f), FileUtils.readFileToString(new File(pathFile2)));
 		} catch (Exception e) {
 			e.printStackTrace();
 			fail();
 		} finally {
-			f.delete();
+			if (f != null){
+				f.delete();
+			}
+			if (f2 != null){
+				f2.delete();
+			}
 		}
 	}
